Allow --seed to make sense when there is more than one batch
If the --seed argument is supplied, we're asking run.py to do
something deterministic based on the given seed. However, we don't
want it to do the same thing for every batch, since that would be a
bit pointless.

This patch slightly changes the meaning of --seed N to "use seed N for
the first batch; seed N+1 for the second batch etc.".

Unfortunately, this still depends on the batch size because we don't
re-seed between tests in a batch. So these two runs will generate
different programs:

    ./run.py ... --steps=gen --iterations=100 --batch-size 5 --seed 123
    ./run.py ... --steps=gen --iterations=100 --batch-size 10 --seed 123

I think we could work around this by using srandom() in the generator
between each test and passing the initial seed as a plusarg (rather
than using -svseed / -sv_seed arguments), but I'll leave this for a
leter patch.